fix(ui): correct stale relative paths after layout and workspaces moves
- layout/AppShell/layout.tsx: Header now at ../Header/Header, not ./Header.
- shared/Modal/ViewInfoModal.tsx: SpacesSwitcher/hooks moved under
  workspaces/SpacesSwitcher/hooks per #19311.
