+++ ui/components/layout/AppShell/layout.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react';
-import Header from './Header';
+import Header from '../Header/Header';
 
 const layoutStyle = { margin: 20, padding: 20, border: '1px solid #DDD' };
